Rework multiple things with FTP in mind
Kinda allows network logons, makes FTP into one of the allowed logon methods.  Adding the _ftp_password_crack vulnerability to the computer in question allows it to happen

A User() object must reside in the computer.users list, with the correct username and password handed to the ftp command when prompted.

Signed-off-by: Misha Turnbull <mishaturnbull@gmail.com>